Assignment 5 completed

+++ Mesh.cpp
@@ -3,7 +3,7 @@
 #include <d3d11.h>
 #include <wrl/client.h>
 
-Mesh::Mesh(const char* name, Vertex* vert, unsigned int totalVertices, unsigned int* indices, unsigned int totalIndices)
+Mesh::Mesh(const char* name, Vertex* vert, size_t totalVertices, unsigned int* indices, size_t totalIndices)
 {
 	// Create a VERTEX BUFFER
 	// - This holds the vertex data of triangles for a single object
@@ -15,7 +15,7 @@ Mesh::Mesh(const char* name, Vertex* vert, unsigned int totalVertices, unsigned 
 		//  - After the buffer is created, this description variable is unnecessary
 		D3D11_BUFFER_DESC vbd = {};
 		vbd.Usage = D3D11_USAGE_IMMUTABLE;	// Will NEVER change
-		vbd.ByteWidth = sizeof(Vertex) * totalVertices;       // number of vertices in the buffer
+		vbd.ByteWidth = sizeof(Vertex) * (UINT)totalVertices;       // number of vertices in the buffer
 		vbd.BindFlags = D3D11_BIND_VERTEX_BUFFER; // Tells Direct3D this is a vertex buffer
 		vbd.CPUAccessFlags = 0;	// Note: We cannot access the data from C++ (this is good)
 		vbd.MiscFlags = 0;
@@ -43,7 +43,7 @@ Mesh::Mesh(const char* name, Vertex* vert, unsigned int totalVertices, unsigned 
 		//  - Bind Flag (used as an index buffer instead of a vertex buffer) 
 		D3D11_BUFFER_DESC ibd = {};
 		ibd.Usage = D3D11_USAGE_IMMUTABLE;	// Will NEVER change
-		ibd.ByteWidth = sizeof(unsigned int) * totalIndices;	// number of indices in the buffer
+		ibd.ByteWidth = sizeof(unsigned int) * (UINT)totalIndices;	// number of indices in the buffer
 		ibd.BindFlags = D3D11_BIND_INDEX_BUFFER;	// Tells Direct3D this is an index buffer
 		ibd.CPUAccessFlags = 0;	// Note: We cannot access the data from C++ (this is good)
 		ibd.MiscFlags = 0;
@@ -59,8 +59,8 @@ Mesh::Mesh(const char* name, Vertex* vert, unsigned int totalVertices, unsigned 
 	}
 
 	this->name = name;
-	this->totalVertices = totalVertices;
-	this->totalIndices = totalIndices;
+	this->totalVertices = (unsigned int)totalVertices;
+	this->totalIndices = (unsigned int)totalIndices;
 }
 
 Mesh::~Mesh()
